Dump sequence schema information after table information.
There is an issue whereby dropping a database and recreating it by loading the schema results in multiple sequences with the same name being created. This happens because the sequences for table ids are created before the tables are created, which means when the table is created it cannot create its id sequence without appending a digit to make it unique. Results in schema noise in version control, and additional unused sequences being created.